refactor(frontend): split vault presentation components
